statTemplate improvements and rename template files
