+++ js/gameStatTemplate.js
