+++ js/rulesTemplate.js
